refractor: refractor basic server
