+++ backend/index.ts
@@ -1,0 +1,2 @@
+
+console.log('Run')
