+++ package.json
@@ -3,10 +3,11 @@
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
     "main": "index.js",
-    "type": "module",
+
     "scripts": {
         "start": "node ./index.js",
         "start:dev": "nodemon ./index.js",
+				"server:dev": "nodemon backend/index.ts",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "author": "Andrei Auchynnikau",
@@ -16,6 +17,10 @@
         "nodemon": "^2.0.16"
     },
     "dependencies": {
+        "ts-node": "^10.9.2",
+        "ts-node-dev": "^2.0.0",
+        "tsx": "^4.7.1",
+        "typescript": "^5.3.3",
         "ws": "^8.8.0"
     }
 }
